@@ -166,12 +166,11 @@
     return nil;
   }
 
-  NSMutableDictionary *userInfo = [NSMutableDictionary new];
-  for (NSErrorUserInfoKey key in error.userInfo) {
-    id value = error.userInfo[key];
-    userInfo[key] = [FIAObjectTranslator encodeNSErrorUserInfo:value];
-  }
-  return @{@"code" : @(error.code), @"domain" : error.domain ?: @"", @"userInfo" : userInfo};
+  return @{
+    @"code" : @(error.code),
+    @"domain" : error.domain ?: @"",
+    @"userInfo" : [FIAObjectTranslator encodeNSErrorUserInfo:error.userInfo]
+  };
 }
 
 + (id)encodeNSErrorUserInfo:(id)value {
@@ -187,6 +186,12 @@
     NSMutableArray *errors = [[NSMutableArray alloc] init];
     for (id error in value) {
       [errors addObject:[FIAObjectTranslator encodeNSErrorUserInfo:error]];
+    }
+    return errors;
+  } else if ([value isKindOfClass:[NSDictionary class]]) {
+    NSMutableDictionary *errors = [[NSMutableDictionary alloc] init];
+    for (id key in value) {
+      errors[key] = [FIAObjectTranslator encodeNSErrorUserInfo:value[key]];
     }
     return errors;
   } else {
