@@ -3,8 +3,9 @@
 // found in the LICENSE file.
 
 import 'dart:js_interop';
-import 'dart:js_interop_unsafe';
+
 import 'package:web/web.dart' as web;
+import 'pkg_web_tweaks.dart';
 
 // Returns the URL to load an asset from this example app as a network source.
 //
@@ -19,40 +20,29 @@ String getUrlForAssetAsNetworkSource(String assetKey) {
       '?raw=true';
 }
 
-extension type Descriptor._(JSObject _) implements JSObject {
-  // May also contain "configurable" and "enumerable" bools.
-  // See: https://developer.mozilla.org/en-US/docs/Web/JavaScript/Reference/Global_Objects/Object/defineProperty#description
-  external factory Descriptor({
-    // bool configurable,
-    // bool enumerable,
-    JSBoolean writable,
-    JSAny value,
-  });
-}
-
-void _defineProperty(
-  Object object,
-  String property,
-  Descriptor description,
-) {
-  (globalContext['Object'] as JSObject?)?.callMethod(
-    'defineProperty'.toJS,
-    object as JSObject,
-    property.toJS,
-    description,
-  );
-}
-
 /// Forces a VideoElement to report "Infinity" duration.
 ///
 /// Uses JS Object.defineProperty to set the value of a readonly property.
-void setInfinityDuration(Object videoElement) {
-  assert(videoElement is web.HTMLVideoElement);
-  _defineProperty(
-      videoElement,
-      'duration',
-      Descriptor(
-        writable: true.toJS,
-        value: double.infinity.toJS,
+void setInfinityDuration(web.HTMLVideoElement element) {
+  DomObject.defineProperty(
+    element,
+    'duration',
+    Descriptor.data(
+      writable: true,
+      value: double.infinity.toJS,
+    ),
+  );
+}
+
+/// Makes the `currentTime` setter throw an exception if used.
+void makeSetCurrentTimeThrow(web.HTMLVideoElement element) {
+  DomObject.defineProperty(
+      element,
+      'currentTime',
+      Descriptor.accessor(
+        set: (JSAny? value) {
+          throw Exception('Unexpected call to currentTime with value: $value');
+        },
+        get: () => 100.toJS,
       ));
 }
